refactor: rewrite code to recreate chrome bug
This commit fixes the alleged bug in chrome which was found out when writing same code in another branch. This now renders the same way in firefox.
What was thought to be a bug isn't actually a bug, the weird rendering got fixed when VSCode  live server was restarted. As to why live server caused that behaviour is still unknown for now.

 Please enter the commit message for your changes. Lines starting